# project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auto CRUD",
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Auto"
                ],
                "summary": "Страница формы создания авто",
                "operationId": "da0f7b9f10223abd24ade878719162db",
                "responses": {
                    "200": {
                        "description": "Html страница"
                    }
                }
            }
        },
        "/autos/": {
            "get": {
                "tags": [
                    "Auto"
                ],
                "summary": "Список автомобилей без подробных характеристик",
                "operationId": "0b70d480240070e5276123447917f1cc",
                "responses": {
                    "200": {
                        "description": "Список",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "code": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "TOYOTA CRESTA"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auto/{auto}": {
            "get": {
                "tags": [
                    "Auto"
                ],
                "summary": "Автомобиль",
                "operationId": "f7f17194ecdb9dc35db180ea5dc09e82",
                "responses": {
                    "200": {
                        "description": "Данные авто",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "model_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "issue": {
                                                    "type": "integer",
                                                    "example": 1980
                                                },
                                                "mileage": {
                                                    "type": "integer",
                                                    "example": 100000
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "red"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-05-02T13:30:14.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-05-02T13:30:14.000000Z"
                                                },
                                                "model": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "CRESTA"
                                                        },
                                                        "mark_id": {
                                                            "type": "integer",
                                                            "example": "1"
                                                        },
                                                        "mark": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "TOYOTA"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Auto"
                ],
                "summary": "Обновление автомобиля",
                "operationId": "87503eaa2ee944eb440a009b040c0d38",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "data": {
                                                "properties": {
                                                    "model_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "issue": {
                                                        "type": "integer",
                                                        "example": 1980
                                                    },
                                                    "mileage": {
                                                        "type": "integer",
                                                        "example": 100000
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "red"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Результат обновления",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auto": {
            "post": {
                "tags": [
                    "Auto"
                ],
                "summary": "Создание автомобиля",
                "operationId": "73706a6026be976a7a3c3b984716e684",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "model_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "issue": {
                                                "type": "integer",
                                                "example": 1980
                                            },
                                            "mileage": {
                                                "type": "integer",
                                                "example": 100000
                                            },
                                            "color": {
                                                "type": "string",
                                                "example": "red"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Данные созданного авто",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "model_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "issue": {
                                                    "type": "integer",
                                                    "example": 1980
                                                },
                                                "mileage": {
                                                    "type": "integer",
                                                    "example": 100000
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "red"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}